+++ CHANGELOG.md
@@ -11,6 +11,8 @@
         - fix the panic for invalid input.
     - `seqkit subseq`:
         - add a new flag `-R/--region-coord` for appending coordinates to sequence ID for `-r/--region`. [#413](https://github.com/shenwei356/seqkit/issues/413)
+    - `seqkit locate`:
+        - add a new flag `-s/--max-len-to-show` to show at most X characters for the search pattern or matched sequences.
     - `seqkit seq`:
         - change the nucleotide color theme. [#412](https://github.com/shenwei356/seqkit/pull/412)
 - [SeqKit v2.5.1](https://github.com/shenwei356/seqkit/releases/tag/v2.5.1) - 2023-08-09